Cancel auto-format on save in vim
In my vimrc, `autocmd BufWritePre` is activate to format lines when
saving. However, it is annoying if you edit an existing code which
is under the control of git and not formatted well. All of codes are
formatted well, but no need to do that usually.

This update is to comment out lines of `autocmd BufWritePre`.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>