Revise python3 installation
Install python3 via apt-get is defined in `Vagrantfile.orig`, but
target version of python can be changed sometimes. So, it is better to
move this installation to an installer script as other packages.

This update is to add `installer/python3.sh` in which targetted
versions, `python3.8` and `python3.9` on the latest Ubuntu 20.04 LTS,
are defined.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>