Move VM params from Vagrantfile
This update is for multi-machine usecase supported in the next updates.

Defining params for VMs can be complexed if the number of VMs is
increased. To avoid to maintain large size of Vagrantfile in such a
situation, add a dedicated file for the param definition `machines.yml`
to be separated from Vagrantfile. This update also add a class to
support loading the params.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>