Set private IP in each VM
If you launch multiple VMs, you need to re-write your IP addresses on
the VMs before running `stack.sh` because this tool does not support
to upload a definition of IPs in `machines.yml` to the VMs. However,
it's convenient to do so from `machines.yml`. This change is to support
the feature.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>