Support multi-distros setup
This update is to introduce multi-distros support which enables to
install OpenStack with devstack not only on ubuntu but also other
distro, mainly targetting RedHat distros.

All of ansible playbooks are moved under a new dedicated directory
named `ubuntu-focal`, and `hosts` and `site.yaml` are also updated for
the change. Playbooks for other distro will be placed in other decicated
directory, such as `roles/centos8` for CentOS8.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>