Remove useless apt pkgs for devstack install
Some libraries installed with devstack script conflict with libs already
installed with apt. `launchpadlib` was only the case of the lib, but 2nd
one appeared, `python3-pyasn1-modules`. It conflicts with another
installation of devstack in which tries to uninstall already existing
lib `pyasn1-modules` at once, but failed to do because `pip uninstall`
cannot resolve a dependency if it's installed via system tool like
`apt`. So, uninstallation is failed as the result.

This update for ansible playbook is to remove `pyans1` with `apt`
command before running devstack script to avoid the conlict. It also
update the name of the playbook because it was only for `launchpadlib`,
but is not for the package now, so renamed as `remove_useless_pkgs.yml`.

The key of this update is to run following command from the playbook.
  $ sudo apt remove python3-pyasn1

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>